% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{grnn_examples}
\alias{grnn_examples}
\title{Examples of a GRNN model}
\usage{
grnn_examples(forecast)
}
\arguments{
\item{forecast}{A \code{grnnForecast} object.}
}
\value{
A matrix including the features and targets of the examples
   associated with the model of a \code{grnnForecast} object.
}
\description{
It shows the examples of the model associated to a
\code{grnnForecast} object.
}
\examples{
pred <- grnn_forecasting(ts(1:8), h = 1, lags = 1:2)
grnn_examples(pred)
}
