% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_origin.R
\name{plot.grnnForecastRO}
\alias{plot.grnnForecastRO}
\title{Plot the prediction for a test set}
\usage{
\method{plot}{grnnForecastRO}(x, h = NULL, ...)
}
\arguments{
\item{x}{the object obtained from a call to \code{\link{rolling_origin}}.}

\item{h}{an integer. The forecasting horizon. If \code{NULL} (the default),
the maximum forecasting horizon of all the test sets is used.}

\item{...}{Other plotting parameters to affect the plot.}
}
\value{
None
}
\description{
It plots the forecast associated with a test set generated with the function
\code{\link{rolling_origin}}.
}
