% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot.grnnForecast}
\alias{autoplot.grnnForecast}
\title{Create a ggplot object from a grnnForecast object}
\usage{
\method{autoplot}{grnnForecast}(forecast, highlight = c("none", "points"))
}
\arguments{
\item{forecast}{The grnnForecast object.}

\item{highlight}{A string value indicating what elements should be
highlighted. Possible values are "none" and "points".}
}
\value{
The ggplot object representing a plotting with the forecast.
}
\description{
It uses a grnnForecast object to create a ggplot object that plots a time
series and its forecast using GRNN regression.
}
\examples{
pred <- grnn_forecasting(USAccDeaths, h = 12, lags = 1:12, sigma = 50)
library(ggplot2)
autoplot(pred)
}
